added dbutils interface
