+++ samples/apps/autogen-studio/autogenstudio/cli.py
@@ -4,7 +4,6 @@ import typer
 import uvicorn
 
 from .version import VERSION
-from .utils.dbutils import DBManager
 
 app = typer.Typer()
