+++ samples/apps/autogen-studio/autogenstudio/utils/__init__.py
@@ -1,2 +1,1 @@
-from .dbutils import *
 from .utils import *
