+++ samples/apps/autogen-studio/autogenstudio/web/app.py
@@ -14,7 +14,8 @@ from ..datamodel import (
     Message,
     Session,
 )
-from ..utils import md5_hash, init_webserver_folders, DBManager, dbutils
+from ..utils import md5_hash, init_webserver_folders
+from ..utils.dbmanager import SqliteDBManager
 
 from ..chatmanager import AutoGenChatManager
 
@@ -51,17 +52,18 @@ api.mount("/files", StaticFiles(directory=folders["files_static_root"], html=Tru
 
 
 db_path = os.path.join(root_file_path, "database.sqlite")
-dbmanager = DBManager(path=db_path)  # manage database operations
+dbmanager = SqliteDBManager()  # manage database operations
+dbmanager.init_db(path=db_path)  # initialize database
 chatmanager = AutoGenChatManager()  # manage calls to autogen
 
 
 @api.post("/messages")
 async def add_message(req: ChatWebRequestModel):
     message = Message(**req.message.dict())
-    user_history = dbutils.get_messages(user_id=message.user_id, session_id=req.message.session_id, dbmanager=dbmanager)
+    user_history = dbmanager.get_messages(user_id=message.user_id, session_id=req.message.session_id)
 
     # save incoming message to db
-    dbutils.create_message(message=message, dbmanager=dbmanager)
+    dbmanager.create_message(message=message)
     user_dir = os.path.join(folders["files_static_root"], "user", md5_hash(message.user_id))
     os.makedirs(user_dir, exist_ok=True)
 
@@ -74,7 +76,7 @@ async def add_message(req: ChatWebRequestModel):
         )
 
         # save assistant response to db
-        dbutils.create_message(message=response_message, dbmanager=dbmanager)
+        dbmanager.create_message(message=response_message)
         response = {
             "status": True,
             "message": response_message.content,
@@ -94,7 +96,7 @@ async def get_messages(user_id: str = None, session_id: str = None):
     if user_id is None:
         raise HTTPException(status_code=400, detail="user_id is required")
     try:
-        user_history = dbutils.get_messages(user_id=user_id, session_id=session_id, dbmanager=dbmanager)
+        user_history = dbmanager.get_messages(user_id=user_id, session_id=session_id)
 
         return {
             "status": True,
@@ -112,7 +114,7 @@ async def get_messages(user_id: str = None, session_id: str = None):
 @api.get("/gallery")
 async def get_gallery_items(gallery_id: str = None):
     try:
-        gallery = dbutils.get_gallery(gallery_id=gallery_id, dbmanager=dbmanager)
+        gallery = dbmanager.get_gallery(gallery_id=gallery_id)
         return {
             "status": True,
             "data": gallery,
@@ -133,7 +135,7 @@ async def get_user_sessions(user_id: str = None):
         raise HTTPException(status_code=400, detail="user_id is required")
 
     try:
-        user_sessions = dbutils.get_sessions(user_id=user_id, dbmanager=dbmanager)
+        user_sessions = dbmanager.get_sessions(user_id=user_id)
 
         return {
             "status": True,
@@ -155,7 +157,7 @@ async def create_user_session(req: DBWebRequestModel):
 
     try:
         session = Session(user_id=req.session.user_id, flow_config=req.session.flow_config)
-        user_sessions = dbutils.create_session(user_id=req.user_id, session=session, dbmanager=dbmanager)
+        user_sessions = dbmanager.create_session(user_id=req.user_id, session=session)
         return {
             "status": True,
             "message": "Session created successfully",
@@ -174,7 +176,7 @@ async def publish_user_session_to_gallery(req: DBWebRequestModel):
     """Create a new session for a user"""
 
     try:
-        gallery_item = dbutils.create_gallery(req.session, tags=req.tags, dbmanager=dbmanager)
+        gallery_item = dbmanager.create_gallery(req.session, tags=req.tags)
         return {
             "status": True,
             "message": "Session successfully published",
@@ -193,7 +195,7 @@ async def delete_user_session(req: DBWebRequestModel):
     """Delete a session for a user"""
 
     try:
-        sessions = dbutils.delete_session(session=req.session, dbmanager=dbmanager)
+        sessions = dbmanager.delete_session(session=req.session)
         return {
             "status": True,
             "message": "Session deleted successfully",
@@ -212,8 +214,8 @@ async def remove_message(req: DeleteMessageWebRequestModel):
     """Delete a message from the database"""
 
     try:
-        messages = dbutils.delete_message(
-            user_id=req.user_id, msg_id=req.msg_id, session_id=req.session_id, dbmanager=dbmanager
+        messages = dbmanager.delete_message(
+            user_id=req.user_id, msg_id=req.msg_id, session_id=req.session_id
         )
         return {
             "status": True,
@@ -231,7 +233,7 @@ async def remove_message(req: DeleteMessageWebRequestModel):
 @api.get("/skills")
 async def get_user_skills(user_id: str):
     try:
-        skills = dbutils.get_skills(user_id, dbmanager=dbmanager)
+        skills = dbmanager.get_skills(user_id)
 
         return {
             "status": True,
@@ -249,7 +251,7 @@ async def get_user_skills(user_id: str):
 @api.post("/skills")
 async def create_user_skills(req: DBWebRequestModel):
     try:
-        skills = dbutils.upsert_skill(skill=req.skill, dbmanager=dbmanager)
+        skills = dbmanager.upsert_skill(skill=req.skill)
 
         return {
             "status": True,
@@ -270,7 +272,7 @@ async def delete_user_skills(req: DBWebRequestModel):
     """Delete a skill for a user"""
 
     try:
-        skills = dbutils.delete_skill(req.skill, dbmanager=dbmanager)
+        skills = dbmanager.delete_skill(req.skill)
 
         return {
             "status": True,
@@ -289,7 +291,7 @@ async def delete_user_skills(req: DBWebRequestModel):
 @api.get("/agents")
 async def get_user_agents(user_id: str):
     try:
-        agents = dbutils.get_agents(user_id, dbmanager=dbmanager)
+        agents = dbmanager.get_agents(user_id)
 
         return {
             "status": True,
@@ -309,7 +311,7 @@ async def create_user_agents(req: DBWebRequestModel):
     """Create a new agent for a user"""
 
     try:
-        agents = dbutils.upsert_agent(agent_flow_spec=req.agent, dbmanager=dbmanager)
+        agents = dbmanager.upsert_agent(agent_flow_spec=req.agent)
 
         return {
             "status": True,
@@ -330,7 +332,7 @@ async def delete_user_agent(req: DBWebRequestModel):
     """Delete an agent for a user"""
 
     try:
-        agents = dbutils.delete_agent(agent=req.agent, dbmanager=dbmanager)
+        agents = dbmanager.delete_agent(agent=req.agent)
 
         return {
             "status": True,
@@ -349,7 +351,7 @@ async def delete_user_agent(req: DBWebRequestModel):
 @api.get("/models")
 async def get_user_models(user_id: str):
     try:
-        models = dbutils.get_models(user_id, dbmanager=dbmanager)
+        models = dbmanager.get_models(user_id)
 
         return {
             "status": True,
@@ -369,7 +371,7 @@ async def create_user_models(req: DBWebRequestModel):
     """Create a new model for a user"""
 
     try:
-        models = dbutils.upsert_model(model=req.model, dbmanager=dbmanager)
+        models = dbmanager.upsert_model(model=req.model)
 
         return {
             "status": True,
@@ -390,7 +392,7 @@ async def delete_user_model(req: DBWebRequestModel):
     """Delete a model for a user"""
 
     try:
-        models = dbutils.delete_model(model=req.model, dbmanager=dbmanager)
+        models = dbmanager.delete_model(model=req.model)
 
         return {
             "status": True,
@@ -409,7 +411,7 @@ async def delete_user_model(req: DBWebRequestModel):
 @api.get("/workflows")
 async def get_user_workflows(user_id: str):
     try:
-        workflows = dbutils.get_workflows(user_id, dbmanager=dbmanager)
+        workflows = dbmanager.get_workflows(user_id)
 
         return {
             "status": True,
@@ -429,7 +431,7 @@ async def create_user_workflow(req: DBWebRequestModel):
     """Create a new workflow for a user"""
 
     try:
-        workflow = dbutils.upsert_workflow(workflow=req.workflow, dbmanager=dbmanager)
+        workflow = dbmanager.upsert_workflow(workflow=req.workflow)
         return {
             "status": True,
             "message": "Workflow created successfully",
@@ -449,7 +451,7 @@ async def delete_user_workflow(req: DBWebRequestModel):
     """Delete a workflow for a user"""
 
     try:
-        workflow = dbutils.delete_workflow(workflow=req.workflow, dbmanager=dbmanager)
+        workflow = dbmanager.delete_workflow(workflow=req.workflow)
         return {
             "status": True,
             "message": "Workflow deleted successfully",
